moved router to outside of notificationsContext
