--- desktopApp/reactApp/src/contexts/ComputationContext.jsx
@@ -1,6 +1,0 @@
-import React from "react";
-
-// set the defaults
-const ComputationContext = React.createContext(null);
-
-export default ComputationContext;
